refactor(core): move runtime-agnostic media capability predicates to core/media
Capability predicates (`isMediaPauseCapable`, `isMediaSeekCapable`, etc.)
duck-type on the capability interfaces in `core/media/types` and don't
depend on any DOM types, so they belong alongside the types they guard.

DOM-side feature consumers now import these predicates directly from
`core/media/predicate`. `dom/media/predicate.ts` keeps DOM-specific
guards like `isQuerySelectorAllCapable`.

This unblocks non-DOM consumers (e.g., react-native, future shims) from
importing the predicates without pulling DOM types.
