+++ packages/core/src/dom/store/features/live.ts
@@ -1,8 +1,7 @@
 import { listen } from '@videojs/utils/dom';
-
+import { isMediaLiveCapable } from '../../../core/media/predicate';
 import type { MediaLiveState } from '../../../core/media/state';
 import { definePlayerFeature } from '../../feature';
-import { isMediaLiveCapable } from '../../media/predicate';
 
 /**
  * Player feature exposing `liveEdgeStart` and `targetLiveWindow` in store
