+++ packages/core/src/dom/store/features/playback-rate.ts
@@ -1,8 +1,7 @@
 import { listen } from '@videojs/utils/dom';
-
+import { isMediaPlaybackRateCapable } from '../../../core/media/predicate';
 import type { MediaPlaybackRateState } from '../../../core/media/state';
 import { definePlayerFeature } from '../../feature';
-import { isMediaPlaybackRateCapable } from '../../media/predicate';
 
 const DEFAULT_RATES: readonly number[] = [0.2, 0.5, 0.7, 1, 1.2, 1.5, 1.7, 2];
